:art: Renamed common_functions.py
