+++ .flaskenv
@@ -2,4 +2,3 @@ FLASK_APP=app.py
 FLASK_ENV=development
 FLASK_RUN_PORT=5001
 FLASK_RUN_HOST=localhost
-DATABASE_URL="postgresql://adam:paswird1@127.0.0.1:5432/audit"
