+++ tasks.py
@@ -37,4 +37,6 @@ def virtualenv(c):
 def make_test_db(c, database_host="localhost"):
     """Create a clean database for testing"""
     c.run(f"dropdb -h {database_host} --if-exists fsd_app_store_test")
+    print(stylize("fsd_app_store_test db dropped...", ECHO_STYLE))
     c.run(f"createdb -h {database_host} fsd_app_store_test")
+    print(stylize("fsd_app_store_test db created...", ECHO_STYLE))
